pyproject.toml
[tool.poetry]
name = "dbt-rpc"
version = "0.0.0"
description = "dbt RPC server"
authors = ["dbt Labs <info@dbtlabs.com>"]
readme = "README.rst"
homepage = "https://github.com/dbt-labs/dbt-rpc-server"
repository = "https://github.com/dbt-labs/dbt-rpc-server"
classifiers = [
    "Development Status :: 1 - Planning",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]

[tool.poetry.dependencies]
python = "^3.6.3"
dbt-core = "^0.20.0"
# local editable install
# dbt-core = { path = "../dbt/core", develop = true }
# you will still need to install adapters
dbt-postgres = "^0.20.0"
fastapi = "^0.68.1"
uvicorn = "^0.15.0"
ipdb = "^0.13.9"
websockets = "^9.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
coverage = {extras = ["toml"], version = "^5.4"}
safety = "^1.10.3"
mypy = "^0.910"
typeguard = "^2.12.1"
xdoctest = {extras = ["colors"], version = "^0.15.5"}
sphinx = "^4.1.1"
sphinx-autobuild = "^2021.3.14"
pre-commit = "^2.13.0"
flake8 = "^3.9.2"
black = "^20.8b1"
flake8-bandit = "^2.1.2"
flake8-bugbear = "^21.4.3"
flake8-docstrings = "^1.6.0"
flake8-rst-docstrings = "^0.2.3"
pep8-naming = "^0.12.0"
darglint = "^1.8.0"
reorder-python-imports = "^2.5.0"
pre-commit-hooks = "^4.0.1"
sphinx-rtd-theme = "^0.5.2"
sphinx-click = "^3.0.1"
Pygments = "^2.9.0"

[tool.poetry.scripts]
dbt-rpc = "dbt_rpc.__main__:main"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["dbt_rpc"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.mypy]
strict = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
